Introduced new vertex / fragment shaders for shaded models.
Added a global light.
Fixed an issue with model data indexing.
Added rendering code for a test model.

+++ Src/Rendering/Objects/Model.cpp
@@ -37,7 +37,7 @@ void Model::ParseModelData(Managers::Model::ModelData* p_Data)
 			}
 		}
 
-		if (s_Index != ~0)
+		if (s_Index != (uint16_t) ~0)
 		{
 			m_Indices.push_back(s_Index);
 			continue;
